float cant de element

+++ presentation/routers/ofertas_router.py
@@ -152,7 +152,7 @@ async def delete_oferta(
 async def add_elemento_to_oferta(
     oferta_id: str = Path(..., description="ID de la oferta"),
     categoria: str = Form(..., description="Categoría del elemento"),
-    cantidad: int = Form(..., description="Cantidad del elemento (mayor a 0)"),
+    cantidad: float = Form(..., description="Cantidad del elemento (mayor a 0)"),
     descripcion: Optional[str] = Form(None, description="Descripción del elemento"),
     foto: Optional[UploadFile] = File(None, description="Foto del elemento"),
     oferta_service: OfertaService = Depends(get_oferta_service)
